+++ openVag/GraphLayout.cpp
@@ -2,6 +2,7 @@
 
 #include <list>
 #include <map>
+#include <unordered_map>
 #include <utility>
 #include <tuple>
 #include <cassert>
@@ -11,11 +12,41 @@
 
 #include "imgui.h"
 
+class BiMapConstLayerToOwnerLayer
+{
+public:
+    void insertConstToOwner(std::shared_ptr<Layer> constLayer, std::shared_ptr<Layer> ownerLayer) {
+        mapConstLayerToOwnerLayer[constLayer] = ownerLayer;
+        mapOwnerLayerToVecConstLayer[ownerLayer].push_back(constLayer);
+    }
+    std::vector<std::shared_ptr<Layer>> getVecConstLayerForOwner(std::shared_ptr<Layer> ownerLayer) const {
+        auto it = mapOwnerLayerToVecConstLayer.find(ownerLayer);
+        if (it == mapOwnerLayerToVecConstLayer.end()) return {};
+        return it->second;
+    }
+    std::shared_ptr<Layer> getOwnerLayerForConstLayer(std::shared_ptr<Layer> constLayer) const {
+        auto it = mapConstLayerToOwnerLayer.find(constLayer);
+        if (it == mapConstLayerToOwnerLayer.end()) return nullptr;
+        return it->second;
+    }
+    std::vector<std::shared_ptr<Layer>> getAllConstLayers() {
+        std::vector<std::shared_ptr<Layer>> res;
+        for (const auto& mapVal : mapConstLayerToOwnerLayer) {
+            res.push_back(mapVal.first);
+        }
+        return res;
+    }
+
+private:
+    std::map<std::shared_ptr<Layer>, std::shared_ptr<Layer>> mapConstLayerToOwnerLayer;
+    std::map<std::shared_ptr<Layer>, std::vector<std::shared_ptr<Layer>>> mapOwnerLayerToVecConstLayer;
+};
+
 class Row {
 public:
     std::list<std::shared_ptr<Layer>> vecLayer;
-    ImVec2 getSize(float horizontalSpacing) const;
-    void layoutLayers(float horizontalSpacing, float leftmargin, float horizontalOffset, float verticalOffset) const;
+    ImVec2 getSize(float horizontalSpacing, float verticalSpacing, const BiMapConstLayerToOwnerLayer& biMapConstLayerToOwnerLayer) const;
+    void layoutLayers(float horizontalSpacing, float verticalSpacing, float leftmargin, float horizontalOffset, float verticalOffset, const BiMapConstLayerToOwnerLayer& biMapConstLayerToOwnerLayer) const;
 };
 
 class Graph {
@@ -23,43 +54,88 @@ public:
     Graph(const std::shared_ptr<Layer>& startLayer);
     std::set<std::shared_ptr<Layer>> getAllLayers();
     void layoutLayers(float horizontalSpacing, float verticalSpacing, ImVec2 startPos);
-    ImVec2 getSize(float horizontalSpacing, float verticalSpacing) const;
+    ImVec2 getSize(float horizontalSpacing, float verticalSpacing, const BiMapConstLayerToOwnerLayer& biMapConstLayerToOwnerLayer) const;
+    BiMapConstLayerToOwnerLayer biMapConstLayerToOwnerLayer;
 private:
+    std::list<Row> listRow;
     std::list<Row>::iterator getItRow(std::shared_ptr<Layer>);
     std::list<Row>::iterator makePrevRowAvailable(std::list<Row>::iterator it);
     std::list<Row>::iterator makeNextRowAvailable(std::list<Row>::iterator it);
-    std::list<Row> listRow;
+    std::list<Row>::iterator makeParameterLayerRowAvailable();
+    std::list<Row>::iterator makeResultLayerRowAvailable();
+    std::list<Row>::iterator getDedicatedRow(std::shared_ptr<Layer> layer);
+    std::list<Row>::iterator itInputLayerRow = listRow.end();
+    std::list<Row>::iterator itResultLayerRow = listRow.end();
+
 };
 
 Graph::Graph(const std::shared_ptr<Layer>& startLayer) {
+
     std::deque<std::shared_ptr<Layer>> dequeProcLayer = { startLayer };
     std::set<std::shared_ptr<Layer>> setSeenLayers;
-    listRow.push_back(Row());
-    listRow.front().vecLayer.push_back(startLayer);
+    {
+        auto startLayerTemp = startLayer;
+        if (std::string(startLayerTemp->getType()) == "Const") {
+            auto setOutputLayer = startLayerTemp->getOutputLayers();
+            if (setOutputLayer.size()) {
+                dequeProcLayer.clear();
+                dequeProcLayer.push_back(*setOutputLayer.begin());
+                startLayerTemp = *setOutputLayer.begin();
+            }
+        }
+        auto itDedicatedRow = getDedicatedRow(startLayerTemp);
+        auto itInsertRow = itDedicatedRow != listRow.end() ? itDedicatedRow : listRow.insert(listRow.begin(), Row());
+        itInsertRow->vecLayer.push_back(startLayerTemp);
+    }
     while (dequeProcLayer.size()) {
         auto procLayer = dequeProcLayer.front(); dequeProcLayer.pop_front();
         if (setSeenLayers.find(procLayer) != setSeenLayers.end()) continue;
 
         setSeenLayers.insert(procLayer);
-        auto itRow = getItRow(procLayer);
+        auto skipLayer = [&](const std::shared_ptr<Layer>& layer) {
+                if (setSeenLayers.find(layer) != setSeenLayers.end()) return true;
+                if (std::find(dequeProcLayer.begin(), dequeProcLayer.end(), layer) != dequeProcLayer.end()) return true;
+                return false;
+            };
+
+        auto itRow = [&]() {
+            auto activeLayer = procLayer;
+            if (std::string(activeLayer->getType()) == "Const") {
+                activeLayer = biMapConstLayerToOwnerLayer.getOwnerLayerForConstLayer(activeLayer);
+                assert(activeLayer != nullptr);
+            }
+            return getItRow(activeLayer);
+            }();
+        assert(itRow != listRow.end());
         {
             auto setOutputLayer = procLayer->getOutputLayers();
-            for (auto& layer : setSeenLayers) setOutputLayer.erase(layer);
-            for (auto& layer : dequeProcLayer) setOutputLayer.erase(layer);
-            if (setOutputLayer.size()) {
-                auto itNextRow = makeNextRowAvailable(itRow);
-                itNextRow->vecLayer.insert(itNextRow->vecLayer.end(), setOutputLayer.begin(), setOutputLayer.end());
-                dequeProcLayer.insert(dequeProcLayer.end(), setOutputLayer.begin(), setOutputLayer.end());
+            for (const auto& layer : setOutputLayer) {
+                if (skipLayer(layer)) continue;
+                dequeProcLayer.push_back(layer);
+                if (std::string(layer->getType()) == "Const") {
+                    biMapConstLayerToOwnerLayer.insertConstToOwner(layer, procLayer);
+                    continue;
+                }
+                auto itDedicatedRow = getDedicatedRow(layer);
+                auto itInsertRow = itDedicatedRow != listRow.end() ? itDedicatedRow : makeNextRowAvailable(itRow);
+                itInsertRow->vecLayer.push_back(layer);
+                
             }
         }
         {
             auto setInputLayer = procLayer->getInputLayers();
-            for (auto& layer : setSeenLayers) setInputLayer.erase(layer);
-            for (auto& layer : dequeProcLayer) setInputLayer.erase(layer);
-            if (setInputLayer.size()) {
-                auto itPrevRow = makePrevRowAvailable(itRow);
-                itPrevRow->vecLayer.insert(itPrevRow->vecLayer.end(), setInputLayer.begin(), setInputLayer.end());
-                dequeProcLayer.insert(dequeProcLayer.end(), setInputLayer.begin(), setInputLayer.end());
+            for (const auto& layer : setInputLayer) {
+                if (skipLayer(layer)) continue;
+                dequeProcLayer.push_back(layer);
+                if (std::string(layer->getType()) == "Const") {
+                    if (biMapConstLayerToOwnerLayer.getOwnerLayerForConstLayer(layer) == nullptr) {
+                        biMapConstLayerToOwnerLayer.insertConstToOwner(layer, procLayer);
+                    }
+                    continue;
+                }
+                auto itDedicatedRow = getDedicatedRow(layer);
+                auto itInsertRow = itDedicatedRow != listRow.end() ? itDedicatedRow : makePrevRowAvailable(itRow);
+                itInsertRow->vecLayer.push_back(layer);
             }
         }
     }
@@ -69,30 +145,32 @@ std::set<std::shared_ptr<Layer>> Graph::getAllLayers()
 {
     std::set<std::shared_ptr<Layer>> setLayers;
     for (const auto& row : listRow) { setLayers.insert(row.vecLayer.begin(), row.vecLayer.end()); }
+    auto vecConstLayers = biMapConstLayerToOwnerLayer.getAllConstLayers();
+    setLayers.insert(vecConstLayers.begin(), vecConstLayers.end());
     return setLayers;
 }
 
 void Graph::layoutLayers(float horizontalSpacing, float verticalSpacing, ImVec2 startPos)
 {
-    auto graphSize = getSize(horizontalSpacing, verticalSpacing);
+    auto graphSize = getSize(horizontalSpacing, verticalSpacing, biMapConstLayerToOwnerLayer);
     float verticalOffset = 0;
     for (const auto& row : listRow) {
-        auto rowSize = row.getSize(horizontalSpacing);
+        auto rowSize = row.getSize(horizontalSpacing, verticalSpacing, biMapConstLayerToOwnerLayer);
         auto leftMargin = (graphSize.x - rowSize.x) / 2;
-        row.layoutLayers(horizontalSpacing, leftMargin, startPos.x, startPos.y + verticalOffset);
+        row.layoutLayers(horizontalSpacing, verticalSpacing, leftMargin, startPos.x, startPos.y + verticalOffset, biMapConstLayerToOwnerLayer);
         verticalOffset += verticalSpacing + rowSize.y;
     }
 }
 
-ImVec2 Graph::getSize(float horizontalSpacing, float vericalSpacing) const
+ImVec2 Graph::getSize(float horizontalSpacing, float verticalSpacing, const BiMapConstLayerToOwnerLayer& biMapConstLayerToOwnerLayer) const
 {
     ImVec2 graphSize;
     for (auto& row : listRow) {
-        auto rowSize = row.getSize(horizontalSpacing);
+        auto rowSize = row.getSize(horizontalSpacing, verticalSpacing, biMapConstLayerToOwnerLayer);
         graphSize.x = std::max(graphSize.x, rowSize.x);
-        graphSize.y += rowSize.y + vericalSpacing;
+        graphSize.y += rowSize.y + verticalSpacing;
     }
-    graphSize.y -= vericalSpacing; //Last row does not needd verticalSpacing
+    graphSize.y -= verticalSpacing; //Last row does not needd verticalSpacing
     return graphSize;
 }
 
@@ -109,18 +187,48 @@ std::list<Row>::iterator Graph::getItRow(std::shared_ptr<Layer> searchLayer)
 
 std::list<Row>::iterator Graph::makePrevRowAvailable(std::list<Row>::iterator it)
 {
+    if (it == itInputLayerRow) {
+        return listRow.insert(std::next(itInputLayerRow), Row());
+    }
     if (it == listRow.begin()) {
         return listRow.insert(listRow.begin(), Row());
     }
-    return --it;
+    if (std::prev(it) == itInputLayerRow) {
+        return listRow.insert(std::next(itInputLayerRow), Row());
+    }
+    return std::prev(it);
 }
 
 std::list<Row>::iterator Graph::makeNextRowAvailable(std::list<Row>::iterator it)
 {
     auto nextIt = ++it;
 
-    if (nextIt != listRow.end()) { return nextIt; }
-    return listRow.insert(listRow.end(), Row());
+    if (nextIt != itResultLayerRow) { return nextIt; }
+    return listRow.insert(itResultLayerRow, Row());
+}
+
+std::list<Row>::iterator Graph::makeParameterLayerRowAvailable()
+{
+    if (itInputLayerRow != listRow.end()) return itInputLayerRow;
+    itInputLayerRow = listRow.insert(listRow.begin(), Row());
+    return itInputLayerRow;
+}
+
+std::list<Row>::iterator Graph::makeResultLayerRowAvailable()
+{
+    if (itResultLayerRow != listRow.end()) return itResultLayerRow;
+    itResultLayerRow = listRow.insert(listRow.end(), Row());
+    return itResultLayerRow;
+}
+
+std::list<Row>::iterator Graph::getDedicatedRow(std::shared_ptr<Layer> layer)
+{
+    const auto layerType = std::string(layer->getType());
+    if (layerType == "Parameter")
+        return makeParameterLayerRowAvailable();
+    if (layerType == "Result")
+        return makeResultLayerRowAvailable();
+    return listRow.end();
 }
 
 void GraphLayout::layoutNodes(std::set<std::shared_ptr<Layer>, LayerIDLess> layers, ImVec2 startPos) {
@@ -135,14 +243,27 @@ void GraphLayout::layoutNodes(std::set<std::shared_ptr<Layer>, LayerIDLess> laye
 
     for (auto& graph : vecGraph) {
         graph.layoutLayers(horizontalSpacing, verticalSpacing, startPos);
-        startPos.x += (graph.getSize(horizontalSpacing, verticalSpacing)).x + horizontalSpacing;
+        startPos.x += (graph.getSize(horizontalSpacing, verticalSpacing, graph.biMapConstLayerToOwnerLayer)).x + horizontalSpacing;
     }
 }
 
-ImVec2 Row::getSize(float horizontalSpacing) const {
+static ImVec2 getNodeSize(const std::shared_ptr<Layer>& layer, float horizontalSpacing, float verticalSpacing, const BiMapConstLayerToOwnerLayer& biMapConstLayerToOwnerLayer) {
+    auto ownerNodeSize = ax::NodeEditor::GetNodeSize(layer->getId());
+    auto vecConst = biMapConstLayerToOwnerLayer.getVecConstLayerForOwner(layer);
+    ImVec2 vecConstSize(0.0, 0.0);
+    for (const auto& constLayer : vecConst) {
+        auto constLayerSize = getNodeSize(constLayer, horizontalSpacing, verticalSpacing, biMapConstLayerToOwnerLayer);
+        vecConstSize.x += constLayerSize.x;
+        vecConstSize.y = std::max(vecConstSize.y, constLayerSize.y);
+    }
+    if (vecConst.size()) { vecConstSize.x += (vecConst.size() - 1) * horizontalSpacing; }
+    return ImVec2(std::max(vecConstSize.x, ownerNodeSize.x), vecConst.size() ? vecConstSize.y + verticalSpacing + ownerNodeSize.y : ownerNodeSize.y);
+}
+
+ImVec2 Row::getSize(float horizontalSpacing, float verticalSpacing, const BiMapConstLayerToOwnerLayer& biMapConstLayerToOwnerLayer) const {
     ImVec2 rowSize;
     for (const auto& layer : vecLayer) {
-        auto nodeSize = ax::NodeEditor::GetNodeSize(layer->getId());
+        auto nodeSize = getNodeSize(layer, horizontalSpacing, verticalSpacing, biMapConstLayerToOwnerLayer);
         rowSize.x += nodeSize.x;
         rowSize.y = std::max(rowSize.y, nodeSize.y);
     }
@@ -150,12 +271,28 @@ ImVec2 Row::getSize(float horizontalSpacing) const {
     return rowSize;
 }
 
-void Row::layoutLayers(float horizontalSpacing, float leftmargin, float horizontalOffset, float verticalOffset) const
+static void SetLayerPosition(std::shared_ptr<Layer> layer, ImVec2 pos, float horizontalSpacing, float verticalSpacing, const BiMapConstLayerToOwnerLayer& biMapConstLayerToOwnerLayer)
+{
+    auto wholeNodeSize = getNodeSize(layer, horizontalSpacing, verticalSpacing, biMapConstLayerToOwnerLayer);
+    auto ownerNodeSize = ax::NodeEditor::GetNodeSize(layer->getId());
+    ImVec2 actualOwnerNodePos = pos;
+    actualOwnerNodePos.x = pos.x + (wholeNodeSize.x - ownerNodeSize.x) / 2;
+    actualOwnerNodePos.y = pos.y + (wholeNodeSize.y - ownerNodeSize.y);
+    ax::NodeEditor::SetNodePosition(layer->getId(), actualOwnerNodePos);
+    auto vecConstLayer = biMapConstLayerToOwnerLayer.getVecConstLayerForOwner(layer);
+    ImVec2 constLayerPos(pos.x, pos.y);
+    for (const auto& constLayer : vecConstLayer) {
+        SetLayerPosition(constLayer, constLayerPos, horizontalSpacing, verticalSpacing, biMapConstLayerToOwnerLayer);
+        constLayerPos.x += horizontalSpacing + getNodeSize(constLayer, horizontalSpacing, verticalSpacing, biMapConstLayerToOwnerLayer).x;
+    }
+}
+
+void Row::layoutLayers(float horizontalSpacing, float verticalSpacing, float leftmargin, float horizontalOffset, float verticalOffset, const BiMapConstLayerToOwnerLayer& biMapConstLayerToOwnerLayer) const
 {
     auto xOffset = leftmargin + horizontalOffset;
     for (const auto& layer : vecLayer) {
         ImVec2 pos(xOffset, verticalOffset);
-        ax::NodeEditor::SetNodePosition(layer->getId(), pos);
-        xOffset += ax::NodeEditor::GetNodeSize(layer->getId()).x + horizontalSpacing;
+        SetLayerPosition(layer, pos, horizontalSpacing, verticalSpacing, biMapConstLayerToOwnerLayer);
+        xOffset += getNodeSize(layer, horizontalSpacing, verticalSpacing, biMapConstLayerToOwnerLayer).x + horizontalSpacing;
     }
 }