+++ openVag/IRModel.h.LOCAL.h
@@ -70,7 +70,6 @@ public:
 
     virtual std::set<std::shared_ptr<Edge>, EdgeIDLess> getSetEdge() const = 0;
     std::vector<std::string> getVecDim();
-    void setVecDim(std::vector<std::string> vecDim, std::vector<size_t> vecDimPos);
     void setAttributes(std::vector<std::pair<std::string, std::string>> vecAttribute);
 private:
     virtual void modifyEdgeAttributesAfterIdChange(std::shared_ptr<Edge> edge) = 0;