+++ changelog.md
@@ -35,3 +35,20 @@
 * Fixed redo command
 * Removed edge arrows
 
+## [v1.1.1]
+
+### Features
+
+* Properties panel is now displaying port dimensions
+* Validation message for edges that are connecting two ports with different dimensions
+* "Const" Layers are disappearing when they are not selected or not connected to a selected node
+* Ports that are connected to "Const" layer are colored in gray.
+
+### Changes
+
+* Properties panel does no longer expand layer properties when more than two layers are selected
+* Properties panel has "goto" buttons when displaying edge properties that allows to navigate canvas to connecting layers
+* Fixed crash when undo and redo after deleting a port
+* When creating graph layout the first row is dedicated to layers of type: "Parameter" and the last row is dedicated to layers of type "Result"
+* Layers of type "Const" are treated as part of the layer they are connected to.
+